+++ docs/changelog.md
@@ -1,7 +1,20 @@
-## [2.1.0] - January 19 2021
+## [2.2.0] - February 12 2021
+### Added
+- ``IconController.clearIconOnSpawn(icon)``
+
+### Improved
+- The cleanup process when ``icon:destroy()`` is called
+
+
+
+--------
+
+## [2.1.0] - February 2 2021
 ### Added
 - ``icon:bindEvent(iconEventName, eventFunction)``
 - ``icon:unbindEvent(iconEventName)``
+
+
 
 --------